refactor pjoject struct
