--- src/game/tile/TileColor.java
@@ -1,9 +1,0 @@
-package game.tile;
-
-public enum TileColor {
-    BLACK("흑"),
-    WHITE("백");
-
-    TileColor(String color) {
-    }
-}
