--- src/game/tile/TileType.java
@@ -1,9 +1,0 @@
-package game.tile;
-
-public enum TileType {
-    JOKER("조커"),
-    NUMBER("숫자");
-
-    TileType(String type) {
-    }
-}
